added routes to navbar
