+++ package.json
@@ -31,6 +31,7 @@
     "react": "^16.13.1",
     "react-bootstrap": "^2.2.0",
     "react-dom": "^16.13.1",
+    "react-router-bootstrap": "^0.25.0",
     "react-router-dom": "^5.3.0",
     "react-scripts": "^3.4.1"
   },
