+++ src/components/App.js
@@ -1,29 +1,14 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 // getAPIHealth is defined in our axios-services directory index.js
 // you can think of that directory as a collection of api adapters
 // where each adapter fetches specific info from our express server's /api route
-import { getAPIHealth } from "../axios-services";
 import "../style/App.css";
 import Navbar from "react-bootstrap/Navbar";
 import Routes from "../Routes";
+import Nav from "react-bootstrap/Nav";
+import { LinkContainer } from "react-router-bootstrap";
 
 const App = () => {
-  const [APIHealth, setAPIHealth] = useState("");
-
-  useEffect(() => {
-    // follow this pattern inside your useEffect calls:
-    // first, create an async function that will wrap your axios service adapter
-    // invoke the adapter, await the response, and set the data
-    const getAPIStatus = async () => {
-      const { healthy } = await getAPIHealth();
-      setAPIHealth(healthy ? "api is up! :D" : "api is down :/");
-    };
-
-    // second, after you've defined your getter above
-    // invoke it immediately after its declaration, inside the useEffect callback
-    getAPIStatus();
-  }, []);
-
   return (
     <div className="App container py-3">
       <Navbar collapseOnSelect bg="light" expand="md" className="mb-3">
@@ -31,10 +16,22 @@ const App = () => {
           Socks4u
         </Navbar.Brand>
         <Navbar.Toggle />
+        <Navbar.Collapse className="justify-content-end">
+          <Nav activeKey={window.location.pathname}></Nav>
+          <Nav>
+            <LinkContainer to="/signup">
+              <Nav.Link href="/signup">Signup</Nav.Link>
+            </LinkContainer>
+            <LinkContainer to="/login">
+              <Nav.Link href="/login">Login</Nav.Link>
+            </LinkContainer>
+            <LinkContainer to="/API">
+              <Nav.Link href="/API">API</Nav.Link>
+            </LinkContainer>
+          </Nav>
+        </Navbar.Collapse>
       </Navbar>
       <Routes />
-      <h1>Hello, World!</h1>
-      <p>API Status: {APIHealth}</p>
     </div>
   );
 };
